refactor: Rename `…/event-detection/event-types/cloudFront.js`
