+++ sdk-js/src/lib/event-detection/event-types/cloud-front.js
