+++ sdk-js/src/lib/event-detection/index.js
@@ -4,7 +4,7 @@ const alexaSkill = require('./event-types/alexa-skill');
 const apiGateway = require('./event-types/api-gateway');
 const apiGatewayV2 = require('./event-types/api-gateway-v2');
 const customAuthorizer = require('./event-types/customAuthorizer');
-const cloudFront = require('./event-types/cloudFront');
+const cloudFront = require('./event-types/cloud-front');
 const cloudwatchEvent = require('./event-types/cloudwatch-event');
 const cloudwatchLogs = require('./event-types/cloudwatch-log');
 const dynamodb = require('./event-types/dynamodb');
